requirements.txt
fastapi
uvicorn
requests
azure-data-tables
python-dotenv
